+++ getsrc.sh
@@ -25,6 +25,7 @@ mkdir -p src
 
 download http://www.repeatmasker.org/rmblast-2.10.0+-x64-linux.tar.gz
 download http://eddylab.org/software/hmmer/hmmer-3.3.tar.gz
+download https://github.com/Benson-Genomics-Lab/TRF/archive/v4.09.1.tar.gz trf-4.09.1.tar.gz
 download http://www.repeatmasker.org/RepeatScout-1.0.6.tar.gz
 download http://www.repeatmasker.org/RepeatModeler/RECON-1.08.tar.gz
 download https://github.com/weizhongli/cdhit/releases/download/V4.8.1/cd-hit-v4.8.1-2019-0228.tar.gz